pyproject.toml
[tool.poetry]
name = "weather-bot"
version = "0.1.0"
description = "A simple and fast Telegram Bot to get weather info"
authors = ["Ledmington <ledmington.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
geopy = "^2.2.0"
python-telegram-bot = "^13.12"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
